test: split context coverage from clawsweeper suite
